+++ libraries/resource_providers/api300/c7000/switch_provider.rb
@@ -14,6 +14,7 @@ module OneviewCookbook
     module C7000
       # Switch API300 C7000 provider
       class SwitchProvider < API200::SwitchProvider
+        include OneviewCookbook::PortActions::UpdatePort
       end
     end
   end