+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## New Features:
 - [#346](https://github.com/HewlettPackard/oneview-chef/issues/346) Add action to reapply configuration of oneview_interconnect API500
+- [#348](https://github.com/HewlettPackard/oneview-chef/issues/348) Add the update_port action in the Switch API300
 
 ## 3.0.0
 Adds API 500 support to the following HPE OneView resources: